+++ finch/processes/wpsio.py
@@ -304,6 +304,18 @@ output_prefix = copy_io(
     abstract="Prefix of the output filename, defaults to the identifier of the process."
 )
 
+csv_precision = LiteralInput(
+    "csv_precision",
+    "Number of decimal places to round to in the CSV output.",
+    abstract=(
+        "Only valid if output_format is CSV. If not set, all decimal places of a 64 bit "
+        "floating precision number are printed. If negative, rounds before the decimal point."
+    ),
+    data_type="integer",
+    min_occurs=0,
+    max_occurs=1
+)
+
 xclim_common_options = [
     check_missing,
     missing_options,